# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "webstie titlte",
        "description": "API documention",
        "termsOfService": "https://example.com",
        "contact": {
            "name": "mahdi",
            "email": "mahdishoorabi@gmail.com",
            "url": "https://example.com"
        },
        "license": {
            "name": "License Name",
            "url": "https://example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/days": {
            "get": {
                "tags": [
                    "days"
                ],
                "summary": "Get all days with routines",
                "operationId": "8f4ac099c7a75bf9f23bbcc1c0645d25",
                "responses": {
                    "200": {
                        "description": "A list of days with routines"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "days"
                ],
                "summary": "Store a new day",
                "operationId": "1f2177b4ca1fe7d65135bcda7ed74d89",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Day created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/days/{id}": {
            "get": {
                "tags": [
                    "days"
                ],
                "summary": "Get a specific day by column and value",
                "operationId": "4a3866e3c3ef7494432a7066c00ed9cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A specific day"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "days"
                ],
                "summary": "Update an existing day",
                "operationId": "3369b716dba2716b6eb5207d992e78c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Day updated successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "days"
                ],
                "summary": "Delete a day",
                "operationId": "7feeaf84f4a57a3ab22acfb8f522ab4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Day deleted successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/goals": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Get all goals",
                "operationId": "d9cb79f325f8eae93bddc30c289e2f76",
                "responses": {
                    "200": {
                        "description": "List of goals"
                    }
                }
            },
            "post": {
                "tags": [
                    "Goals"
                ],
                "summary": "Create a new goal",
                "operationId": "6127de0f82207012c84fda169fc40ff2",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Goal created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/goals/{id}": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Get a specific goal",
                "operationId": "be77e120089c0eca42cfec5bde43dc91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal details"
                    },
                    "404": {
                        "description": "Goal not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Goals"
                ],
                "summary": "Update a goal",
                "operationId": "aa1ece6b002c55429ef6015761ba5f9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Goal updated successfully"
                    },
                    "404": {
                        "description": "Goal not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Goals"
                ],
                "summary": "Delete a goal",
                "operationId": "b816e7dc3db48e5c322c432078425c3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal deleted successfully"
                    },
                    "404": {
                        "description": "Goal not found"
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}": {
            "put": {
                "tags": [
                    "sessions"
                ],
                "summary": "Update an existing session",
                "description": "Updates a session by ID",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Session updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "sessions"
                ],
                "summary": "Delete a session",
                "description": "Deletes a session by ID",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}